+++ Network/Sources/Network/API/TodoAPI/MainTodoAPI.swift
@@ -1,5 +1,5 @@
 //
-//  MainTodoAPI.swift
+//  TodoAPI.swift
 //  
 //
 //  Created by 김지현 on 2022/10/01.
@@ -7,7 +7,9 @@
 
 import Foundation
 
-protocol MainTodoAPIProtocol {
+protocol TodoAPIProtocol {
+
+  // MARK: - todo main view
   func getTodosData(
     completion: @escaping (
       BaseResponseType<MainTodoDTO.Response.MainTodoResponseDTO>?, Error?
@@ -33,16 +35,34 @@ protocol MainTodoAPIProtocol {
     _ id: Int?,
     _ dto: UpdateTodoDTO.ModifyTodo,
     completion: @escaping (BaseResponseType<String>?, Error?
-                          ) -> Void
+    ) -> Void
   )
   func deleteTodo(
     _ todoId: Int,
     completion: @escaping (BaseResponseType<String>?, Error?
-                          ) -> Void
+    ) -> Void
+  )
+
+  // MARK: - 요일별 todo view
+  func getDaysOfWeekTodosData(
+    completion: @escaping (BaseResponseType<ByDayTodoDTO.Response.ByDayTodosResponseDTO>?, Error?
+    ) -> Void
+  )
+
+  // MARK: - 멤버별 todo view
+  func getMemberTodosData(
+    completion: @escaping (
+      BaseResponseType<MemberTodoDTO.Response.MemberTodosResponseDTO>?, Error?
+    ) -> Void
+  )
+
+  func getModifyingTodo(
+    todoID: Int,
+    completion: @escaping (BaseResponseType<UpdateTodoDTO.ModifyTodo>?, Error?) -> Void
   )
 }
 
-public final class MainTodoAPI: APIRequestLoader<MainTodoService>, MainTodoAPIProtocol {
+public final class TodoAPI: APIRequestLoader<TodoService>, TodoAPIProtocol {
 
   public func checkTodo(
     _ todoId: Int,
@@ -86,8 +106,8 @@ public final class MainTodoAPI: APIRequestLoader<MainTodoService>, MainTodoAPIPr
       target: .getTodoSummary(todoId),
       responseData: BaseResponseType<MainTodoDTO.Response.TodoSummaryResponseDTO>.self
     ) { res, err in
-        completion(res, err)
-      }
+      completion(res, err)
+    }
   }
 
   public func updateTodo(
@@ -112,6 +132,48 @@ public final class MainTodoAPI: APIRequestLoader<MainTodoService>, MainTodoAPIPr
     fetchData(
       target: .deleteTodo(todoId),
       responseData: BaseResponseType<String>.self
+    ) { res, err in
+      completion(res, err)
+    }
+  }
+
+  // MARK: - 요일별
+  public func getDaysOfWeekTodosData(completion: @escaping (BaseResponseType<ByDayTodoDTO.Response.ByDayTodosResponseDTO>?, Error?) -> Void) {
+      fetchData(target: .getDaysOfWeekTodos,
+                responseData: BaseResponseType<ByDayTodoDTO.Response.ByDayTodosResponseDTO>.self,
+                isWithInterceptor: true) { res, err in
+          completion(res, err)
+      }
+  }
+
+  public func getOnlyMyTodoData(completion: @escaping (BaseResponseType<ByDayTodoDTO.Response.OnlyMyTodoDTO>?, Error?) -> Void) {
+      fetchData(target: .getOnlyMyTodo,
+                responseData: BaseResponseType<ByDayTodoDTO.Response.OnlyMyTodoDTO>.self,
+                isWithInterceptor: true) { res, err in
+          completion(res, err)
+      }
+  }
+
+  // MARK: - 멤버별
+  public func getModifyingTodo(
+    todoID: Int,
+    completion: @escaping (BaseResponseType<UpdateTodoDTO.ModifyTodo>?, Error?) -> Void
+  ) {
+    fetchData(
+      target: .getModifyTodoID(todoID),
+      responseData: BaseResponseType<UpdateTodoDTO.ModifyTodo>.self) { res, err in
+        completion(res,err)
+      }
+  }
+
+  public func getMemberTodosData(
+    completion: @escaping (
+      BaseResponseType<MemberTodoDTO.Response.MemberTodosResponseDTO>?, Error?) -> Void
+  ) {
+    fetchData(
+      target: .getMemberTodos,
+      responseData: BaseResponseType<MemberTodoDTO.Response.MemberTodosResponseDTO>.self,
+      isWithInterceptor: true
     ) { res, err in
       completion(res, err)
     }
