+++ Network/Sources/Network/Base/NetworkService.swift
@@ -14,12 +14,9 @@ final public class NetworkService {
 
   public let authRepository = AuthAPI(isLogging: true)
   public let mainHomeRepository = MainHomeAPI(isLogging: true)
+  public let todoRepository = TodoAPI(isLogging: true)
   public let roomRepository = RoomAPI(isLogging: true)
   public let ruleRepository = RuleAPI(isLogging: true)
   public let profileRepository = ProfileAPI(isLogging: true)
   public let userRepository = UserAPI(isLogging: true)
-
-  public let mainTodoRepository = MainTodoAPI(isLogging: true)
-  public let memberTodoRepository = MemberTodoAPI(isLogging: true)
-  public let byDayTodoRepository = ByDayTodoAPI(isLogging: true)
 }
